refacotr: 채팅 생성하기 훅구현
